Liste projets clotures et mail contact PO

+++ Projet2Crowdfunding/ViewModels/HomeViewModel.cs
@@ -30,6 +30,6 @@ namespace Projet2Crowdfunding.ViewModels
 
         public int ClosedProjects { get; set; }
 
-        public int AmountTotal { get; set; }
+        public double AmountTotal { get; set; }
     }
 }
